flytt funksjon for beregning av tilkjent ytelse ut i egen generator-fil + flytt relevante tester
