+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/beregning/TilkjentYtelseUtils.kt
@@ -9,19 +9,13 @@ import no.nav.familie.ba.sak.common.inkluderer
 import no.nav.familie.ba.sak.common.sisteDagIInneværendeMåned
 import no.nav.familie.ba.sak.common.toYearMonth
 import no.nav.familie.ba.sak.ekstern.restDomene.RestYtelsePeriode
-import no.nav.familie.ba.sak.kjerne.beregning.UtvidetBarnetrygdUtil.finnUtvidetVilkår
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelseMedEndreteUtbetalinger
 import no.nav.familie.ba.sak.kjerne.beregning.domene.EndretUtbetalingAndelMedAndelerTilkjentYtelse
-import no.nav.familie.ba.sak.kjerne.beregning.domene.InternPeriodeOvergangsstønad
 import no.nav.familie.ba.sak.kjerne.beregning.domene.TilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ba.sak.kjerne.beregning.domene.medEndring
-import no.nav.familie.ba.sak.kjerne.endretutbetaling.domene.Årsak
-import no.nav.familie.ba.sak.kjerne.fagsak.FagsakType
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Person
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonType
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlag
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import no.nav.familie.ba.sak.kjerne.tidslinje.komposisjon.slåSammenLike
 import no.nav.familie.ba.sak.kjerne.tidslinje.månedPeriodeAv
@@ -29,7 +23,6 @@ import no.nav.familie.ba.sak.kjerne.tidslinje.tidspunkt.tilYearMonth
 import no.nav.familie.ba.sak.kjerne.tidslinje.tilTidslinje
 import no.nav.familie.ba.sak.kjerne.tidslinjefamiliefelles.transformasjon.ZipPadding
 import no.nav.familie.ba.sak.kjerne.tidslinjefamiliefelles.transformasjon.zipMedNeste
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkårsvurdering
 import no.nav.familie.tidslinje.Periode
 import no.nav.familie.tidslinje.Tidslinje
 import no.nav.familie.tidslinje.mapVerdi
@@ -38,128 +31,9 @@ import no.nav.familie.tidslinje.utvidelser.filtrerIkkeNull
 import no.nav.familie.tidslinje.utvidelser.kombinerMed
 import no.nav.familie.tidslinje.utvidelser.tilPerioderIkkeNull
 import java.math.BigDecimal
-import java.time.LocalDate
 import java.time.YearMonth
 
 object TilkjentYtelseUtils {
-    fun beregnTilkjentYtelse(
-        vilkårsvurdering: Vilkårsvurdering,
-        personopplysningGrunnlag: PersonopplysningGrunnlag,
-        endretUtbetalingAndeler: List<EndretUtbetalingAndelMedAndelerTilkjentYtelse> = emptyList(),
-        fagsakType: FagsakType,
-        skalBrukeNyVersjonAvOppdaterAndelerMedEndringer: Boolean = true,
-        hentPerioderMedFullOvergangsstønad: (aktør: Aktør) -> List<InternPeriodeOvergangsstønad> = { _ -> emptyList() },
-    ): TilkjentYtelse {
-        val tilkjentYtelse =
-            TilkjentYtelse(
-                behandling = vilkårsvurdering.behandling,
-                opprettetDato = LocalDate.now(),
-                endretDato = LocalDate.now(),
-            )
-
-        val (endretUtbetalingAndelerSøker, endretUtbetalingAndelerBarna) = endretUtbetalingAndeler.partition { it.person?.type == PersonType.SØKER }
-
-        val andelerTilkjentYtelseBarnaUtenEndringer =
-            OrdinærBarnetrygdUtil
-                .beregnAndelerTilkjentYtelseForBarna(
-                    personopplysningGrunnlag = personopplysningGrunnlag,
-                    personResultater = vilkårsvurdering.personResultater,
-                    fagsakType = fagsakType,
-                ).map {
-                    if (it.person.type != PersonType.BARN) throw Feil("Prøver å generere ordinær andel for person av typen ${it.person.type}")
-
-                    AndelTilkjentYtelse(
-                        behandlingId = vilkårsvurdering.behandling.id,
-                        tilkjentYtelse = tilkjentYtelse,
-                        aktør = it.person.aktør,
-                        stønadFom = it.stønadFom,
-                        stønadTom = it.stønadTom,
-                        kalkulertUtbetalingsbeløp = it.beløp,
-                        nasjonaltPeriodebeløp = it.beløp,
-                        type = YtelseType.ORDINÆR_BARNETRYGD,
-                        sats = it.sats,
-                        prosent = it.prosent,
-                    )
-                }
-
-        val barnasAndelerInkludertEtterbetaling3ÅrEller3MndEndringer =
-            if (skalBrukeNyVersjonAvOppdaterAndelerMedEndringer) {
-                oppdaterAndelerMedEndretUtbetalingAndeler(
-                    andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
-                    endretUtbetalingAndeler = endretUtbetalingAndelerBarna.filter { it.årsak in listOf(Årsak.ETTERBETALING_3ÅR, Årsak.ETTERBETALING_3MND) },
-                    tilkjentYtelse = tilkjentYtelse,
-                )
-            } else {
-                oppdaterTilkjentYtelseMedEndretUtbetalingAndelerGammel(
-                    andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
-                    endretUtbetalingAndeler = endretUtbetalingAndelerBarna.filter { it.årsak in listOf(Årsak.ETTERBETALING_3ÅR, Årsak.ETTERBETALING_3MND) },
-                )
-            }
-
-        val andelerTilkjentYtelseUtvidetMedAlleEndringer =
-            UtvidetBarnetrygdUtil.beregnTilkjentYtelseUtvidet(
-                utvidetVilkår = finnUtvidetVilkår(vilkårsvurdering),
-                tilkjentYtelse = tilkjentYtelse,
-                andelerTilkjentYtelseBarnaMedEtterbetaling3ÅrEller3MndEndringer = barnasAndelerInkludertEtterbetaling3ÅrEller3MndEndringer,
-                endretUtbetalingAndelerSøker = endretUtbetalingAndelerSøker,
-                personResultater = vilkårsvurdering.personResultater,
-                skalBrukeNyVersjonAvOppdaterAndelerMedEndringer = skalBrukeNyVersjonAvOppdaterAndelerMedEndringer,
-            )
-
-        val småbarnstilleggErMulig =
-            erSmåbarnstilleggMulig(
-                utvidetAndeler = andelerTilkjentYtelseUtvidetMedAlleEndringer,
-                barnasAndeler = barnasAndelerInkludertEtterbetaling3ÅrEller3MndEndringer,
-            )
-
-        val andelerTilkjentYtelseSmåbarnstillegg =
-            if (småbarnstilleggErMulig) {
-                SmåbarnstilleggBarnetrygdGenerator(
-                    behandlingId = vilkårsvurdering.behandling.id,
-                    tilkjentYtelse = tilkjentYtelse,
-                ).lagSmåbarnstilleggAndeler(
-                    perioderMedFullOvergangsstønad =
-                        hentPerioderMedFullOvergangsstønad(
-                            personopplysningGrunnlag.søker.aktør,
-                        ),
-                    utvidetAndeler = andelerTilkjentYtelseUtvidetMedAlleEndringer,
-                    barnasAndeler = barnasAndelerInkludertEtterbetaling3ÅrEller3MndEndringer,
-                    barnasAktørerOgFødselsdatoer =
-                        personopplysningGrunnlag.barna.map {
-                            Pair(
-                                it.aktør,
-                                it.fødselsdato,
-                            )
-                        },
-                )
-            } else {
-                emptyList()
-            }
-
-        val andelerTilkjentYtelseBarnaMedAlleEndringer =
-            if (skalBrukeNyVersjonAvOppdaterAndelerMedEndringer) {
-                oppdaterAndelerMedEndretUtbetalingAndeler(
-                    andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
-                    endretUtbetalingAndeler = endretUtbetalingAndelerBarna,
-                    tilkjentYtelse = tilkjentYtelse,
-                )
-            } else {
-                oppdaterTilkjentYtelseMedEndretUtbetalingAndelerGammel(
-                    andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
-                    endretUtbetalingAndeler = endretUtbetalingAndelerBarna,
-                )
-            }
-
-        tilkjentYtelse.andelerTilkjentYtelse.addAll(andelerTilkjentYtelseBarnaMedAlleEndringer.map { it.andel } + andelerTilkjentYtelseUtvidetMedAlleEndringer.map { it.andel } + andelerTilkjentYtelseSmåbarnstillegg.map { it.andel })
-
-        return tilkjentYtelse
-    }
-
-    private fun erSmåbarnstilleggMulig(
-        utvidetAndeler: List<AndelTilkjentYtelseMedEndreteUtbetalinger>,
-        barnasAndeler: List<AndelTilkjentYtelseMedEndreteUtbetalinger>,
-    ): Boolean = utvidetAndeler.isNotEmpty() && barnasAndeler.isNotEmpty()
-
     fun oppdaterAndelerMedEndretUtbetalingAndeler(
         andelTilkjentYtelserUtenEndringer: Collection<AndelTilkjentYtelse>,
         endretUtbetalingAndeler: List<EndretUtbetalingAndelMedAndelerTilkjentYtelse>,
